# docs/swagger.yaml
definitions:
  entity.Account:
    properties:
      balance:
        type: number
      currency:
        $ref: '#/definitions/entity.Currency'
      currencyId:
        type: string
      dateTime:
        type: string
      id:
        type: string
      user:
        $ref: '#/definitions/entity.User'
      userId:
        type: string
    type: object
  entity.Currency:
    properties:
      id:
        type: string
      name:
        type: string
      symbol:
        type: string
    type: object
  entity.Operation:
    properties:
      dateTime:
        type: string
      details:
        type: object
      id:
        type: string
      transactions:
        items:
          $ref: '#/definitions/entity.Transaction'
        type: array
      type:
        type: integer
    type: object
  entity.Rate:
    properties:
      currency:
        $ref: '#/definitions/entity.Currency'
      currencyId:
        type: string
      dateTime:
        type: string
      id:
        type: string
      value:
        type: number
    type: object
  entity.Transaction:
    properties:
      account:
        $ref: '#/definitions/entity.Account'
      accountId:
        type: string
      accountRateValue:
        type: number
      currency:
        $ref: '#/definitions/entity.Currency'
      currencyId:
        type: string
      currencyRateValue:
        type: number
      dateTime:
        type: string
      id:
        type: string
      operation:
        $ref: '#/definitions/entity.Operation'
      operationId:
        type: string
      value:
        type: number
    type: object
  entity.User:
    properties:
      city:
        type: string
      country:
        description: 'TODO: make countries list with foreign key'
        type: string
      id:
        type: string
      name:
        type: string
    type: object
  operationRepository.RefillForm:
    properties:
      accountId:
        type: string
      currency:
        type: string
      details:
        description: Details of operation
        properties:
          source:
            type: string
        type: object
      sum:
        type: number
    type: object
  operationRepository.TransferForm:
    properties:
      currency:
        type: string
      details:
        description: Details of operation
        properties:
          comment:
            type: string
        type: object
      from:
        type: string
      sum:
        type: number
      to:
        type: string
    type: object
  rateRepository.CreateForm:
    properties:
      symbol:
        type: string
      value:
        type: number
    type: object
  rateRepository.UploadRatesForm:
    properties:
      rates:
        items:
          $ref: '#/definitions/rateRepository.CreateForm'
        type: array
    type: object
  registerRepository.Form:
    properties:
      accountSymbol:
        type: string
      userCity:
        type: string
      userCountry:
        type: string
      userName:
        type: string
    type: object
  reportRepository.AccountTransactionsReportData:
    properties:
      account:
        $ref: '#/definitions/entity.Account'
      dateTime:
        type: string
      since:
        type: string
      till:
        type: string
      transactions:
        items:
          $ref: '#/definitions/reportRepository.AccountTransactionsReportTransaction'
        type: array
    type: object
  reportRepository.AccountTransactionsReportTransaction:
    properties:
      accountRate:
        type: number
      currency:
        type: string
      currencyRate:
        type: number
      dateTime:
        type: string
      type:
        type: string
      value:
        type: number
    type: object
info:
  contact: {}
  description: This is an assignment
  title: Swagger Example API
  version: "1.0"
paths:
  /operation/refill:
    post:
      consumes:
      - application/json
      operationId: operationRefill
      parameters:
      - description: params
        in: body
        name: website
        required: true
        schema:
          $ref: '#/definitions/operationRepository.RefillForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Operation'
      summary: Refill operation
  /operation/transfer:
    post:
      consumes:
      - application/json
      operationId: operationTransfer
      parameters:
      - description: params
        in: body
        name: website
        required: true
        schema:
          $ref: '#/definitions/operationRepository.TransferForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Operation'
      summary: Transfer operation
  /rate/upload:
    post:
      consumes:
      - application/json
      operationId: rateUpload
      parameters:
      - description: params
        in: body
        name: website
        required: true
        schema:
          $ref: '#/definitions/rateRepository.UploadRatesForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entity.Rate'
            type: array
      summary: Upload new rates
  /reg:
    post:
      consumes:
      - application/json
      operationId: register
      parameters:
      - description: params
        in: body
        name: website
        required: true
        schema:
          $ref: '#/definitions/registerRepository.Form'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Account'
      summary: Creates new user + account
  /report/account-transactions:
    get:
      operationId: reportAccountTransactions
      parameters:
      - description: Date of report (eg 2020-10-16)
        in: query
        name: date
        required: true
        type: string
      - description: Account id
        in: query
        name: accountId
        required: true
        type: string
      - description: file output
        in: query
        name: file
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/reportRepository.AccountTransactionsReportData'
      summary: Generates report
swagger: "2.0"